Separate test cases on tabs from `MainTabBarControllerTests` to prevent leftover main thread work from previous test cases from singletons.
