+++ WooCommerce/WooCommerceTests/Mocks/MockPOSEligibilityChecker.swift
@@ -1,0 +1,11 @@
+import Foundation
+@testable import WooCommerce
+
+final class MockPOSEligibilityChecker: POSEntryPointEligibilityCheckerProtocol {
+    var result: POSEligibilityState = .eligible
+
+    @MainActor
+    func checkEligibility() async -> POSEligibilityState {
+        result
+    }
+}
