+++ WooCommerce/WooCommerceTests/ViewRelated/MainTabBarController+TabsTests.swift
@@ -1,0 +1,257 @@
+import TestKit
+import XCTest
+@testable import WooCommerce
+
+final class MainTabBarController_TabsTests: XCTestCase {
+    override func setUp() {
+        super.setUp()
+        SessionManager.removeTestingDatabase()
+    }
+
+    func test_tab_view_controllers_are_not_empty_after_updating_default_site() throws {
+        // Arrange
+        // Sets mock `FeatureFlagService` before `MainTabBarController` is initialized so that the feature flags are set correctly.
+        let featureFlagService = MockFeatureFlagService()
+        featureFlagService.isFeatureFlagEnabledReturnValue[.pointOfSaleAsATabi1] = false
+
+        let storesManager = MockStoresManager(sessionManager: .makeForTesting())
+
+        guard let tabBarController = UIStoryboard(name: "Main", bundle: nil).instantiateInitialViewController(creator: { coder in
+            return MainTabBarController(coder: coder, featureFlagService: featureFlagService, stores: storesManager)
+        }) else {
+            return
+        }
+
+        // Trigger `viewDidLoad`
+        XCTAssertNotNil(tabBarController.view)
+
+        // Action
+        storesManager.updateDefaultStore(storeID: 980)
+
+        // Assert
+        XCTAssertEqual(tabBarController.viewControllers?.count, 4)
+        assertThat(tabBarController.tabRootViewController(tab: .myStore, isPOSTabVisible: false),
+                   isAnInstanceOf: DashboardViewHostingController.self)
+        assertThat(tabBarController.tabRootViewController(tab: .orders, isPOSTabVisible: false),
+                   isAnInstanceOf: OrdersSplitViewWrapperController.self)
+        assertThat(tabBarController.tabRootViewController(tab: .products, isPOSTabVisible: false),
+                   isAnInstanceOf: ProductsViewController.self)
+
+        let hubMenuNavigationController = try XCTUnwrap(tabBarController.tabRootViewController(tab: .hubMenu, isPOSTabVisible: false) as? UINavigationController)
+        assertThat(hubMenuNavigationController.topViewController,
+                   isAnInstanceOf: HubMenuViewController.self)
+    }
+
+    func test_tab_view_controllers_include_pos_tab_when_pos_is_eligible() throws {
+        // Given
+        let featureFlagService = MockFeatureFlagService()
+        featureFlagService.isFeatureFlagEnabledReturnValue[.pointOfSaleAsATabi1] = true
+
+        let mockPOSEligibilityChecker = MockPOSEligibilityChecker()
+        mockPOSEligibilityChecker.result = .eligible
+
+        let storesManager = MockStoresManager(sessionManager: .makeForTesting())
+
+        guard let tabBarController = UIStoryboard(name: "Main", bundle: nil).instantiateInitialViewController(creator: { coder in
+            return MainTabBarController(coder: coder,
+                                        featureFlagService: featureFlagService,
+                                        stores: storesManager,
+                                        posEligibilityCheckerFactory: { _ in mockPOSEligibilityChecker })
+        }) else {
+            return
+        }
+
+        // Trigger `viewDidLoad`
+        XCTAssertNotNil(tabBarController.view)
+
+        // When
+        storesManager.updateDefaultStore(storeID: 314)
+
+        // Then
+        waitUntil {
+            tabBarController.tabRootViewControllers.count == 5
+        }
+        assertThat(tabBarController.tabRootViewController(tab: .myStore, isPOSTabVisible: true),
+                   isAnInstanceOf: DashboardViewHostingController.self)
+        assertThat(tabBarController.tabRootViewController(tab: .orders, isPOSTabVisible: true),
+                   isAnInstanceOf: OrdersSplitViewWrapperController.self)
+        assertThat(tabBarController.tabRootViewController(tab: .products, isPOSTabVisible: true),
+                   isAnInstanceOf: ProductsViewController.self)
+        assertThat(tabBarController.tabRootViewController(tab: .pointOfSale, isPOSTabVisible: true),
+                   isAnInstanceOf: POSTabViewController.self)
+
+        let hubMenuNavigationController = try XCTUnwrap(tabBarController.tabRootViewController(tab: .hubMenu, isPOSTabVisible: true) as? UINavigationController)
+        assertThat(hubMenuNavigationController.topViewController,
+                   isAnInstanceOf: HubMenuViewController.self)
+    }
+
+    func test_tab_view_controllers_exclude_pos_tab_when_pos_is_not_eligible() throws {
+        // Given
+        let featureFlagService = MockFeatureFlagService()
+        featureFlagService.isFeatureFlagEnabledReturnValue[.pointOfSaleAsATabi1] = true
+
+        let mockPOSEligibilityChecker = MockPOSEligibilityChecker()
+        mockPOSEligibilityChecker.result = .ineligible(reason: .notTablet)
+
+        let storesManager = MockStoresManager(sessionManager: .makeForTesting())
+
+        guard let tabBarController = UIStoryboard(name: "Main", bundle: nil).instantiateInitialViewController(creator: { coder in
+            return MainTabBarController(coder: coder,
+                                        featureFlagService: featureFlagService,
+                                        stores: storesManager,
+                                        posEligibilityCheckerFactory: { _ in mockPOSEligibilityChecker })
+        }) else {
+            return
+        }
+
+        // Trigger `viewDidLoad`
+        XCTAssertNotNil(tabBarController.view)
+
+        // When
+        storesManager.updateDefaultStore(storeID: 707)
+
+        // Then
+        waitUntil {
+            tabBarController.tabRootViewControllers.count == 4
+        }
+        assertThat(tabBarController.tabRootViewController(tab: .myStore, isPOSTabVisible: false),
+                   isAnInstanceOf: DashboardViewHostingController.self)
+        assertThat(tabBarController.tabRootViewController(tab: .orders, isPOSTabVisible: false),
+                   isAnInstanceOf: OrdersSplitViewWrapperController.self)
+        assertThat(tabBarController.tabRootViewController(tab: .products, isPOSTabVisible: false),
+                   isAnInstanceOf: ProductsViewController.self)
+
+        let hubMenuNavigationController = try XCTUnwrap(tabBarController.tabRootViewController(tab: .hubMenu, isPOSTabVisible: false) as? UINavigationController)
+        assertThat(hubMenuNavigationController.topViewController,
+                   isAnInstanceOf: HubMenuViewController.self)
+    }
+
+    func test_tab_view_controllers_exclude_pos_tab_when_feature_flag_is_disabled() throws {
+        // Given
+        let featureFlagService = MockFeatureFlagService()
+        featureFlagService.isFeatureFlagEnabledReturnValue[.pointOfSaleAsATabi1] = false
+
+        let mockPOSEligibilityChecker = MockPOSEligibilityChecker()
+        mockPOSEligibilityChecker.result = .eligible
+
+        let storesManager = MockStoresManager(sessionManager: .makeForTesting())
+
+        guard let tabBarController = UIStoryboard(name: "Main", bundle: nil).instantiateInitialViewController(creator: { coder in
+            return MainTabBarController(coder: coder,
+                                        featureFlagService: featureFlagService,
+                                        stores: storesManager,
+                                        posEligibilityCheckerFactory: { _ in mockPOSEligibilityChecker })
+        }) else {
+            return
+        }
+
+        // Trigger `viewDidLoad`
+        XCTAssertNotNil(tabBarController.view)
+
+        // When
+        storesManager.updateDefaultStore(storeID: 802)
+
+        // Then
+        XCTAssertEqual(tabBarController.viewControllers?.count, 4)
+        assertThat(tabBarController.tabRootViewController(tab: .myStore, isPOSTabVisible: false),
+                   isAnInstanceOf: DashboardViewHostingController.self)
+        assertThat(tabBarController.tabRootViewController(tab: .orders, isPOSTabVisible: false),
+                   isAnInstanceOf: OrdersSplitViewWrapperController.self)
+        assertThat(tabBarController.tabRootViewController(tab: .products, isPOSTabVisible: false),
+                   isAnInstanceOf: ProductsViewController.self)
+
+        let hubMenuNavigationController = try XCTUnwrap(tabBarController.tabRootViewController(tab: .hubMenu, isPOSTabVisible: false) as? UINavigationController)
+        assertThat(hubMenuNavigationController.topViewController,
+                   isAnInstanceOf: HubMenuViewController.self)
+    }
+
+    func test_tab_view_controllers_do_not_change_when_pos_eligibility_changes() throws {
+        // Given
+        let featureFlagService = MockFeatureFlagService()
+        featureFlagService.isFeatureFlagEnabledReturnValue[.pointOfSaleAsATabi1] = true
+
+        let mockPOSEligibilityChecker = MockPOSEligibilityChecker()
+        mockPOSEligibilityChecker.result = .ineligible(reason: .featureFlagDisabled)
+
+        let storesManager = MockStoresManager(sessionManager: .makeForTesting())
+
+        guard let tabBarController = UIStoryboard(name: "Main", bundle: nil).instantiateInitialViewController(creator: { coder in
+            return MainTabBarController(coder: coder,
+                                        featureFlagService: featureFlagService,
+                                        stores: storesManager,
+                                        posEligibilityCheckerFactory: { _ in mockPOSEligibilityChecker })
+        }) else {
+            return
+        }
+
+        // Trigger `viewDidLoad`
+        XCTAssertNotNil(tabBarController.view)
+
+        // When
+        storesManager.updateDefaultStore(storeID: 303)
+
+        // Then initial state
+        waitUntil {
+            tabBarController.tabRootViewControllers.count == 4
+        }
+
+        // When - change POS eligibility
+        mockPOSEligibilityChecker.result = .eligible
+
+        // Then tabs remain the same
+        XCTAssertEqual(tabBarController.tabRootViewControllers.count, 4)
+    }
+
+    func test_tab_root_viewControllers_are_replaced_after_updating_to_a_different_site() throws {
+        // Arrange
+        let stores = MockStoresManager(sessionManager: .makeForTesting())
+        guard let tabBarController = UIStoryboard(name: "Main", bundle: nil).instantiateInitialViewController(creator: { coder in
+            return MainTabBarController(coder: coder, featureFlagService: MockFeatureFlagService(), stores: stores)
+        }) else {
+            return
+        }
+
+        // Trigger `viewDidLoad`
+        XCTAssertNotNil(tabBarController.view)
+
+        // Action
+        stores.updateDefaultStore(storeID: 134)
+        let viewControllersBeforeSiteChange = tabBarController.tabRootViewControllers
+        stores.updateDefaultStore(storeID: 630)
+        let viewControllersAfterSiteChange = tabBarController.tabRootViewControllers
+
+        // Assert
+        XCTAssertEqual(viewControllersBeforeSiteChange.count, viewControllersAfterSiteChange.count)
+        XCTAssertNotEqual(viewControllersBeforeSiteChange[WooTab.myStore.visibleIndex(isPOSTabVisible: false)],
+                          viewControllersAfterSiteChange[WooTab.myStore.visibleIndex(isPOSTabVisible: false)])
+        XCTAssertNotEqual(viewControllersBeforeSiteChange[WooTab.orders.visibleIndex(isPOSTabVisible: false)],
+                          viewControllersAfterSiteChange[WooTab.orders.visibleIndex(isPOSTabVisible: false)])
+        XCTAssertNotEqual(viewControllersBeforeSiteChange[WooTab.products.visibleIndex(isPOSTabVisible: false)],
+                          viewControllersAfterSiteChange[WooTab.products.visibleIndex(isPOSTabVisible: false)])
+        XCTAssertNotEqual(viewControllersBeforeSiteChange[WooTab.hubMenu.visibleIndex(isPOSTabVisible: false)],
+                          viewControllersAfterSiteChange[WooTab.hubMenu.visibleIndex(isPOSTabVisible: false)])
+    }
+
+    func test_tab_view_controllers_stay_the_same_after_updating_to_the_same_site() throws {
+        // Arrange
+        let stores = MockStoresManager(sessionManager: .makeForTesting())
+        guard let tabBarController = UIStoryboard(name: "Main", bundle: nil).instantiateInitialViewController(creator: { coder in
+            return MainTabBarController(coder: coder, stores: stores)
+        }) else {
+            return
+        }
+
+        // Trigger `viewDidLoad`
+        XCTAssertNotNil(tabBarController.view)
+
+        // Action
+        let siteID: Int64 = 610
+        stores.updateDefaultStore(storeID: siteID)
+        let viewControllersBeforeSiteChange = try XCTUnwrap(tabBarController.viewControllers)
+        stores.updateDefaultStore(storeID: siteID)
+        let viewControllersAfterSiteChange = try XCTUnwrap(tabBarController.viewControllers)
+
+        // Assert
+        XCTAssertEqual(viewControllersBeforeSiteChange, viewControllersAfterSiteChange)
+    }
+}
